requirements.txt
torch
torchvision
opencv-python
Pillow
numpy
scipy
pyautogui
tqdm
tensorflow
